# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ONEPIC-Backend",
    "description": "Make Node.js",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/",
  "tags": [{
    "name": "users",
    "description": "회원 가입"
  }],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "회원 가입",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "aa@aa.com"
              },
              "password": {
                "type": "string",
                "example": "aa"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          }
        ],
        "summary": "회원 탈퇴",
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "회원 정보 수정",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          },
          {
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "aa@aa.com"
              },
              "password": {
                "type": "string",
                "example": "aa"
              },
              "nickName": {
                "type": "string",
                "example": "aa"
              }
            }
          }
        }
      ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "유저 조회",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                },
                "result": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "description": "유저 리스트",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 12
                      },
                      "nickName": {
                        "type": "string",
                        "example": "aa"
                      },
                      "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTIsIm5pY2tOYW1lIjpudWxsLCJlbWFpbCI6ImFhQGFhLmNvbSIsImlhdCI6MTU0MzA3MzU4NCwiZXhwIjoxNTQzNjc4Mzg0LCJpc3MiOiJPTkVQSUMiLCJzdWIiOiJ1c2VySW5mbyJ9.8lKBePdYO7Wv3bkAyAUUGfxbZwzZk7JNvuMx351Kfpg"
                      },
                      "password": {
                        "type": "string",
                        "example": "aa"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2018-11-24T15:16:33.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2018-11-24T15:34:34.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/signIn": {
      "post": {
        "tags": [
          "signIn"
        ],
        "summary": "로그인",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "aa@aa.com"
              },
              "password": {
                "type": "string",
                "example": "aa"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                },
                "result": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "signIn"
        ],
        "summary": "로그아웃",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      }
    },
    "/boards": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "게시글",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
         
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          },

          {
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
                     
            "properties": {
              "title": {
                "type": "string",
                "example": "파리여행"
              },
              "createdAt": {
                "type": "string",
                "example": "2018-11-24T15:16:33.000Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2018-11-24T15:34:34.000Z"
              },
              "date": {
                "type": "string",
                "example": "2018-11-24T15:34:34.000Z"
              },
              "content":{
                "type": "string",
                "example": "감성파뤼이"
              },
              "userId":{
                "type": "int",
                "example": "1"
              },
              "location":{
                "type": "string",
                "example": "paris"
              },

              "lon":{
                "type": "float",
                "example": 11.111
              },
              "lat":{
                "type": "float",
                "example": 22.222
              },
              "share":{
                "type": "boolean",
                "example": 0
              },
              "imgUrl":{
                "type": "string",
                "example": "kkkkk.kkk"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "boards"
        ],
        "summary": "게시물 수정 ",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          },

          {
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "example": "파리여행"
              },
              "date": {
                "type": "string",
                "example": "20181124"
              },
              "content":{
                "type": "string",
                "example": "감성파뤼이"
              },
              "userId":{
                "type": "int",
                "example": "1"
              },
              "location":{
                "type": "string",
                "example": "paris"
              },

              "lon":{
                "type": "float",
                "example": 11.111
              },
              "lat":{
                "type": "float",
                "example": 22.222
              },
              "share":{
                "type": "boolean",
                "example": 0
              },
              "imgUrl":{
                "type": "string",
                "example": "kkkkk.kkk"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "boards"
        ],
        "parameters": [
          {
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "게시글 조회",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "API 인증 키",
            "default": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Nywibmlja05hbWUiOm51bGwsImVtYWlsIjoiYWFAYWEuY29tIiwiaWF0IjoxNTQzMDY5NTczLCJleHAiOjE1NDM2NzQzNzMsImlzcyI6Ik9ORVBJQyIsInN1YiI6InVzZXJJbmZvIn0.yOCYZq4pEH0wVJciToRkl8W8kcRwufMXW2sUEI8dCks",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                },
                "error": {
                  "type": "string",
                  "example": null
                },
                "result": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "description": "게시물 리스트",
                    "properties": {
                      "title": {
                        "type": "string",
                        "example": "파리여행"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2018-11-24T15:16:33.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2018-11-24T15:34:34.000Z"
                      },
                      "date": {
                        "type": "timestamp",
                        "example": "20181124"
                      },
                      "content":{
                        "type": "string",
                        "example": "감성파뤼이"
                      },
                      "userId":{
                        "type": "int",
                        "example": "1"
                      },
                      "location":{
                        "type": "string",
                        "example": "paris"
                      },
        
                      "lon":{
                        "type": "float",
                        "example": 11.111
                      },
                      "lat":{
                        "type": "float",
                        "example": 22.222
                      },
                      "share":{
                        "type": "boolean",
                        "example": 0
                      },
                      "imgUrl":{
                        "type": "string",
                        "example": "kkkkk.kkk"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}